feat: #1416; endpoint; scores-lib; player-links;
