+++ src/lib/components/page/league/top-players/TopPlayers-Loader.svelte
