+++ src/lib/components/page/league/top-players/loaders/_Placeholder_Right_Row.svelte
